Move extra categories registration from menu.py
